# table-serve-backend/openapi.yaml
openapi: 3.1.0
info:
  title: Table Serve API
  description: |
    REST API for the Table Serve platform - an NFC-based restaurant menu and ordering system.

    Three access levels:
    - **Super Admin** - Platform administrators (Table Serve company). Requires master password to register.
    - **Admin** - Restaurant owners and managers managing their menu, tables, and analytics.
    - **Customer** - Public-facing API for NFC-scanned menu viewing and order placement.
  version: 1.0.0
  contact:
    name: Table Serve Support
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.tableserve.app
    description: Production
...
paths:
...
  /api/auth/sign-up/email:
    post:
      tags: [Auth]
      summary: Sign up with email
      operationId: signUpEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Signed up successfully
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/admin/register:
    post:
      tags: [Admin]
      summary: Register restaurant admin account
      operationId: registerAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAdminRequest'
      responses:
        '201':
          description: Account and organization created
        '409':
          description: Slug already taken
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin dashboard stats
      operationId: getAdminDashboard
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Dashboard data
  /api/admin/profile:
    get:
      tags: [Admin]
      summary: Get organization profile
      operationId: getAdminProfile
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Organization profile
    patch:
      tags: [Admin]
      summary: Update organization profile and customizations
      operationId: updateAdminProfile
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationProfileRequest'
      responses:
        '200':
          description: Profile updated
  /api/admin/categories:
    get:
      tags: [Admin]
      summary: List menu categories
      operationId: listCategories
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: List of categories
    post:
      tags: [Admin]
      summary: Create menu category
      operationId: createCategory
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuCategoryRequest'
      responses:
        '201':
          description: Category created
  /api/admin/categories/{id}:
    patch:
      tags: [Admin]
      summary: Update menu category
      operationId: updateCategory
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Category updated
    delete:
      tags: [Admin]
      summary: Delete menu category
      operationId: deleteCategory
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Category deleted
  /api/admin/menu:
    get:
      tags: [Admin]
      summary: List menu items
      operationId: listMenuItems
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: List of menu items
    post:
      tags: [Admin]
      summary: Create menu item
      operationId: createMenuItem
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemRequest'
      responses:
        '201':
          description: Menu item created
        '403':
          description: Plan limit reached
  /api/admin/menu/{id}:
    patch:
      tags: [Admin]
      summary: Update menu item
      operationId: updateMenuItem
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Menu item updated
    delete:
      tags: [Admin]
      summary: Deactivate menu item
      operationId: deleteMenuItem
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Menu item deactivated
  /api/admin/tables:
    get:
      tags: [Admin]
      summary: List restaurant tables
      operationId: listTables
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: List of tables
    post:
      tags: [Admin]
      summary: Create restaurant table (generates NFC token)
      operationId: createTable
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '201':
          description: Table created with NFC token
  /api/admin/tables/{id}:
    patch:
      tags: [Admin]
      summary: Update table
      operationId: updateTable
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Table updated
    delete:
      tags: [Admin]
      summary: Delete table
      operationId: deleteTable
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Table deleted
  /api/admin/orders:
    get:
      tags: [Admin]
      summary: List orders
      operationId: listOrders
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, served, cancelled]
      responses:
        '200':
          description: List of orders
  /api/admin/orders/{id}:
    get:
      tags: [Admin]
      summary: Get order details
      operationId: getOrder
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Order details
  /api/admin/orders/{id}/status:
    patch:
      tags: [Admin]
      summary: Update order status
      operationId: updateOrderStatus
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, served, cancelled]
      responses:
        '200':
          description: Order status updated
  /api/admin/analytics:
    get:
      tags: [Admin]
      summary: Restaurant analytics
      operationId: getAdminAnalytics
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          description: Analytics data
  /api/admin/members:
    get:
      tags: [Admin]
      summary: List organization members
      operationId: listMembers
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: List of members

  # ─── Customer ──────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    OrgIdHeader:
      name: x-organization-id
      in: header
      required: false
      description: Organization ID context (not required if session has active organization)
      schema:
        type: string
    PeriodQuery:
      name: period
      in: query
      schema:
        type: string
        enum: [today, week, month, quarter, year, custom]
        default: month
  responses:
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    RegisterAdminRequest:
      type: object
      required: [name, email, password, organizationName, organizationSlug]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        organizationName:
          type: string
        organizationSlug:
          type: string
          pattern: '^[a-z0-9-]+$'
...
    UpdateOrganizationProfileRequest:
      type: object
      properties:
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
        description:
          type: string
        primaryColor:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
        secondaryColor:
          type: string
        accentColor:
          type: string
        fontFamily:
          type: string
          enum: [Inter, Roboto, Poppins, 'Playfair Display', Lato, Montserrat]
        menuLayout:
          type: string
          enum: [grid, list]
        showCalories:
          type: boolean
        showAllergens:
          type: boolean
        showPreparationTime:
          type: boolean
        showSpiceLevel:
          type: boolean
        currencySymbol:
          type: string
        currencyCode:
          type: string
          minLength: 3
          maxLength: 3
        taxRate:
          type: number
        serviceChargeRate:
          type: number
        welcomeMessage:
          type: string
        footerText:
          type: string
    CreateMenuCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
    CreateMenuItemRequest:
      type: object
      required: [name, price]
      properties:
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        isChefSpecial:
          type: boolean
        isVegetarian:
          type: boolean
        isVegan:
          type: boolean
        isGlutenFree:
          type: boolean
        spiceLevel:
          type: string
          enum: [none, mild, medium, hot]
        allergens:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        preparationTime:
          type: integer
        calories:
          type: integer
        sortOrder:
          type: integer
...